{fix} fixed env that had leaks when calledafter env was modified

+++ include/builtin.h
@@ -6,7 +6,7 @@
 /*   By: ldevoude <ldevoude@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/10 14:55:15 by ldevoude          #+#    #+#             */
-/*   Updated: 2025/05/29 16:23:56 by ldevoude         ###   ########lyon.fr   */
+/*   Updated: 2025/06/15 10:45:24 by ldevoude         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	is_newline(t_command_exec *node, int *i, bool *newline);
 int	echo_builtin(t_command_exec *node, bool newline, int i);
 
 //env.c
-int env_builtin(t_command_exec *node, t_utils *utils, int i);
+int env_builtin(t_command_exec *node, t_utils *utils, size_t i);
 
 //exit.c
 unsigned int exit_builtin(t_command_exec *node, t_utils *utils);
